redesign MultiModeDrtModule - aggregator of DrtMode-related modules
